Allow literal # in argument text by escaping
Hi,

inspired by [a recent TeX.sx question][1] this allows delimited macros
with `#` in the parameter text by doubling the `#`.
It isn't needed that often but it is useful if you try to analyze the
parameter text of a TeX macro using TeX code. Also it is completely
backward-compatible (`##` always triggers an error without this) and
it is very easy to implement.

For example:

\def\test#1###2{(#1)[#2]}
\test hello#{world}

would result in (hello)[world]

Best regards,
Marcel Krüger

1: https://tex.stackexchange.com/questions/491296/can-i-have-a-delimited-macro-with-a-literal-in-the-parameter-text